Adjust contextualLabel for type validation problem reports
The contextual messages was calculated as part of the console rendering.
